requirements.txt
# requirements.txt
torch>=1.10.0
transformers>=4.18.0
datasets>=2.0.0
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.60.0
wandb>=0.12.0
pyyaml>=6.0